chore: update example create side panel
