--- examples/carbon-for-ibm-products/CreateSidePanel/src/_carbon-settings.scss
@@ -1,12 +1,0 @@
-@import '@carbon/type/scss/type';
-@import '@carbon/themes/scss/themes';
-@import '@carbon/layout/scss/layout';
-@import '@carbon/motion/scss/motion';
-@import 'carbon-components/src/globals/scss/vars';
-@import 'carbon-components/src/globals/scss/mixins';
-
-$feature-flags: (
-  ui-shell: true,
-  grid-columns-16: true,
-  enable-css-custom-properties: true,
-);
