+++ examples/carbon-for-ibm-products/CreateSidePanel/src/_carbon.scss
@@ -1,35 +1,6 @@
-@import './carbon-settings';
+$feature-flags: (
+  enable-css-custom-properties: true,
+);
 
-// Feature flags
-$css--font-face: true;
-$css--plex: true;
-
-// Global styles
-@import '@carbon/type/scss/reset';
-@import 'carbon-components/scss/globals/scss/css--font-face';
-@import 'carbon-components/scss/globals/grid/grid';
-$carbon--theme: $carbon--theme--white;
-
-// Carbon components
-@import 'carbon-components/scss/components/button/button';
-@import 'carbon-components/scss/components/link/link';
-@import 'carbon-components/scss/components/number-input/number-input';
-@import 'carbon-components/scss/components/text-input/text-input';
-@import 'carbon-components/scss/components/dropdown/dropdown';
-@import 'carbon-components/scss/components/ui-shell/ui-shell';
-
-.carbon-theme--white {
-  @include carbon--theme($carbon--theme--white, true);
-}
-
-.carbon-theme--g10 {
-  @include carbon--theme($carbon--theme--g10, true);
-}
-
-.carbon-theme--g90 {
-  @include carbon--theme($carbon--theme--g90, true);
-}
-
-.carbon-theme--g100 {
-  @include carbon--theme($carbon--theme--g100, true);
-}
+@import '@carbon/elements/scss';
+@include carbon--theme();
