--- examples/carbon-for-ibm-products/CreateSidePanel/src/_ibm-cloud-and-cognitive.scss
@@ -1,8 +1,0 @@
-@import '@carbon/ibm-cloud-cognitive/scss/components/CreateSidePanel/index';
-
-.ccs-theme-dropdown {
-  position: fixed;
-  bottom: $spacing-05;
-  left: $spacing-05;
-  min-width: 11.5rem;
-}
